ivy: add floating-point values (dev branch)
Not much interesting here yet but one can do square roots.

This change is on the dev branch because it depends on the new math/big.Float
type which is not only not in 1.4 but is changing rapidly.

What works: basic arithmetic and square roots, printing, vectors and matrix
operations.

There are also predefined high-precision constants e and pi.

Once 1.5 lands, this work will be much futher along and will be integrated
into the main branch.