ivy: multline operator definitions
Syntax:

For one line, the same as before

	def op x = x

For multiple lines, put a newline after the '=' and finish with a blank line:

	def op x =
		x = x*x
		x

Also make the printing look like this for multi-expression entries (even
if they were entered as a single line; we can't tell the difference).

Document the new syntax.

A big TODO now done, and rather neatly if I do say so myself.