ivy: fix program presentation
Make sure that the numbers, chars, etc. printed in the program
are valid input again. This required adding a new method, ProgString,
to provide alternate formatting for program output. For instance,
Char('a').String() prints a, but Char('a').ProgString() prints 'a'.

Also add a configuration for the output, so the test can capture
the output of )op etc.