ivy: disallow ops and vars with the same name
Prepatory to being able to dump the workspace. If x can be both
an op and a variable, what is the meaning of
	x 3
Is it an op invocation or a vector? Remove the ambiguity.

Op definitions are parsed based on their current definitions, so
there are still complications that must be resolved to dump the
workspace, at least to do so textually, which is the goal.

Also fix an unimportant bug in the implementation of assignment.