ivy: handle recursive function definitions
First, make self-recursive functions parse correctly, so

op a n = a n

parses as a recursive invocation rather than as a vector.
Such functions don't work but at least they parse right.
Warn that they won't work, but accept them.

Second, get the output for )save right for mutually
recursive operators by emitting declarations. This
(as well as the warning mentioned above) requires a
way to compute the references of an op.