ivy: improve test coverage, change logical not to "not" from "~"
Add a number of tests to reduce the red in the coverage report from
	go test --coverprofile cover.out -coverpkg ./...
Most of what's uncovered is either error checking that won't trigger
or debugging that isn't enabled in the tests. Overall now at 82.1%
(not that I care much about the actual number).

Change logical not's spelling from "~" to "not". The former is the APL
spelling but in ivy all the other logical operators are words (and, or,
etc.) so "not" is more locallly consistent.