ivy: refactor function and context out of parse
Move the definition of Function (nee function) and Context
(nee execContext) out of the parse package and into its own
exec package. There is more to do, but this does clean up a
lot of the relationships in the parser. Why would execution
context be inside the parse package?

Part of a series of refactorings to enable user-defined
ops to be more generally powerful.