ivy: major cleanup to remove global variables
There were global Context and Config vars in several packages,
with SetConfig etc. functions to control them. Get rid of them by
making Config reachable through Context and pluming Context
everywhere. This cleans up the code a lot and eanbles further
cleanups to come.

One odd result is that the String methods for Value are no longer
useful, since they cannot access the Context unless we put Context
into Value. That would make sense but would make integers much
bigger and would break the embeddings of BigFloat and BigInt,
which in turn would mess up the code.  So instead we just add an
Sprint that takes a context and use it. String is still there but not
reachable by execution of user code.