ivy/parse: fix multiple assignment
In a statement like a=b=1, before this CL the expression b=1 was
wrapped as an Assignment, which got the right value but meant that
if we printed
	a
we got nothing, as supressing printing is what Assignment is for.
So if we have a chained assignment, strip away the wrappers for
the right hand side. Only the top-level assignment should be
wrapped.