ivy: break the mutual dependence between value.Binary and context.EvalBinary
They were mutually recursive but don't need to be. To fix this, we
need to change the parser to call value.Binary instead of
context.EvalBinary. Then we can remove the extra evaluates
in context.EvalBinary and the call to value.Binary within. Now
value.Binary calls context.EvalBinary, but the reverse does not
happen. This will simplify the general function evaluation that
is coming.

Same thing applies to Unary.

Assignments were still an issue, as their evaluation in context.
EvalBinary was patching over incorrect wrappings of a value
in Assignment in cases like a=b=1 or iota a=3. So tidy that
up too, generalizing the code from the previous commit and
using it when evaluating any expression, unary or binary.