parse: if assignment is of a scalar, it can be an element of a vector
This is the first part of fixing the problem handling
	3 x = 1
which should set x to 1 and return 3 1.

This problem is that isScalar did not recognize the assignment
as holding a scalar value.

The parser is still wrong so we need the parentheses:
	3 (x = 1)
but without this change, even that didn't work.