value: make assignment-as-expression work everywhere
In an earlier CL we introduced an Assignment type in the parser that
wraps an expression created by assignment. This serves to suppress
printing of such values, but it causes problems in a number of places
that do type switches to see what they are holding.

Resolve this by introducing a trivial new method to Value: Inner,
which just returns the value. But for the embedded Value in an
Assignment, it pokes through to the inner true value, giving the
right type for the expression.

While we're here, tidy up a few things and implement an odd detail
of APL that allows an assignment to a vector element make the
variable available anywhere else in the vector:

	(x=3) x
		3 3
	y (y=4)
		4 4