ivy: implement assignment to vector element
Also add a Rank() int method to value.Value, which tidies up
some things.

Matrix assignmnt is not done; to do that we need a Matrix to be
passed by pointer, not value. (Vectors work because they are
slices; Matrix is a struct0. That shouldn't be too hard to do but
should be a separate change.

If that works, we could also move some of the logic for
assignment into the value directory.