ivy: implement unary transpose
Binary (general) transpose is very odd, but shouldn't be too hard
now that the groundwork is laid.

An interesting challenge, writing a single routine to transpose a
matrix of arbitrary rank and dimension. Solution is to notice
that if we think of iterating over the elements we get an index
sequence 0,0,0 0,0,1, 0,0,2, 0,1,0 and so on, in little-endian
order. The transposition is made by iterating in big-endian
order instead: 0,0,0, 1,0,0, 2,0,0, 0,1,0. Then it's easy.

General transpose can be done by generalizing the endianness.