ivy: add conditional evaluation
Based on a suggestion by Alan Davies (glxxyz on GitHub), and borrowing
heavily from Dyalog's guards, implement conditional evaluation. The
new construct looks like a binary operator ":", but is of lower
precedence than any other operator. It is enabled only inside
user-defined functions, and has the property that if the left-hand
operand is non-zero, the function returns the right-hand operand.
Otherwise, the right-hand operand is not evaluated and execution
continues.

It is an error if the left-hand operand is not a scalar.

This has been a long time coming, so thanks to Alan for planting
the seed, life for giving me a free day to think about how to do
it, and especially to GitHub user jojomoore2007, who found a way
to write a recursive function that is so amazing and yet abusive
it pushed me over the edge.

Example from testdata/function.ivy:

	# Recursive factorial.
	op fac n =
		n <= 1 : 1
		n * fac n - 1

	# Recursivd gcd.
	op a gcd b =
		a == b: a
		a > b: b gcd a-b
		a gcd b-a

	fac 10
		3628800
	1562 gcd fac 11
		22

Fixes #65
Update #28