ivy: move the )demo special command inside ivy
Instead of making it a separate program, stitch it in to the binary itself,
meaning that the demo can be run without a Go compiler. The idea was
to make it easier for mobile users to see the demo, (see
https://github.com/golang/go/issues/48694), but that may be a
false hope as I/O on mobile does everything possible to avoid the
Unix idea of stdin and stdout.

Still, it does clean up the way the demo works, and allows us to
remove the runner and explain about it.

Use embedding to interpolate the script into the code. We could use
go generate to avoid needign such a modern feature of Go, but
meh.