demo: make it cleaner on mobile platforms
Add a flag to the config so the mobile wrapper can report it's on
a mobile platform, and use this to give instructions for )demo on
mobile. Tell the user in that case to use the Demo button, which
works fine on Android but needs to be implemented on iOS,
but it should be easy.

Also refactor the demo runner a bit so that it prints instructions
from demo.ivy before accepting user input.