mobile: fix the formatting problems
Fix for https://github.com/golang/go/issues/48693

The output from ivy on mobile gets passed to an HTML renderer.
There was protection for < and > but not all things. And that
protection stops me running a full HTML escaper.

So instead, we escape that stuff that empirically causes trouble,
convert spaces to non-breaking spaces so they don't collapse,
and turn tabs (which outside of user-written code only appear
at the beginning of lines) into 4 spaces for display on mobile.

Because the standard tests lazily leaned on the mobile
framework to run the tests, that laziness had to be fixed.
Now there is much cleaner separation between mobile and
the rest, and the mobile code gets more help from ivy
in the task of running scripts.

Mostly much cleaner. Not much new code, it's mostly
deleting, rearranging, and tweaking, but it's a substantial
rearrangement of the tests.