ivy: add @ operator modifier
The @ operator modifier is a bit like APL's each/map.
Applied to the left or right of an operator, it produces an operator
that applies to each element of its left or right argument and then
returns an array of those results.

A simple example:

	1 2 3 ,@ 4 5 6
		(1 2 3 4) (1 2 3 5) (1 2 3 6)

	1 2 3 @, 4 5 6
		(1 4 5 6) (2 4 5 6) (3 4 5 6)

@f@ is like o.f (outer product) when used with the number of @s
matching the number of dimensions of the left and right arguments.

	1 2 3 @,@ 4 5 6
		(1 4) (1 5) (1 6)
		(2 4) (2 5) (2 6)
		(3 4) (3 5) (3 6)

	1 2 3 o., 4 5 6
		(1 4) (1 5) (1 6)
		(2 4) (2 5) (2 6)
		(3 4) (3 5) (3 6)

	1 o., 4 5 6
		(1 4) (1 5) (1 6)

	1 ,@ 4 5 6
		(1 4) (1 5) (1 6)

A more common use of @ is with user-defined operators.
For example, "shift" shifts a matrix a certain number of positions
up and to the left.

	op d shift m = transp d[2] rot transp d[1] rot m

	x = 2 4 4 rho ".....\\/../\\.....", "...../\\..\\/....."; x
		....
		.\/.
		./\.
		....

		....
		./\.
		.\/.
		....

	1 1 shift x[1]
		\/..
		/\..
		....
		....

shift@ applies to each element in the first axis of x:

	mix 1 1 shift@ x
	\/..
	/\..
	....
	....

	/\..
	\/..
	....
	....

d @shift x[1] applies each direction to x[1]:

	d=(4 2 rho 1 1 1 -1 -1 1 -1 -1); d
		 1  1
		 1 -1
		-1  1
		-1 -1

	mix d @shift x[1]
		\/..
		/\..
		....
		....

		....
		....
		\/..
		/\..

		..\/
		../\
		....
		....

		....
		....
		..\/
		../\

d @shift@ x does both:

	mix d @shift@ x
		[1 * * *]:
		  \/..
		  /\..
		  ....
		  ....

		  /\..
		  \/..
		  ....
		  ....

		[2 * * *]:
		  ....
		  ....
		  \/..
		  /\..

		  ....
		  ....
		  /\..
		  \/..

		[3 * * *]:
		  ..\/
		  ../\
		  ....
		  ....

		  ../\
		  ..\/
		  ....
		  ....

		[4 * * *]:
		  ....
		  ....
		  ..\/
		  ../\

		  ....
		  ....
		  ../\
		  ..\/

@f and f@ roughly correspond to K's "each left" and "each right".
https://code.kx.com/q/ref/maps/#each-left-and-each-right