ivy: in binary iota return highest index of left argument when not found
When origin is set to 0, currently iota will return 0 for missing items
and the first items. When origin is set to 1, iota will also return for
missing items, but these can't be used as an index.
This change makes it so itota returns the max index of the left
argument. This value can be used to index in an array.

     1 -1 0["()" iota "(1 + (4 * 5))"]
         1 0 0 0 0 1 0 0 0 0 0 -1 -1