ivy: finish generalized indexing, fix shape of index result
Generalized indexing like 'ab'[3 3 rho 1 2] handled a few special cases.
Make it handle all cases, simplifying the code. Also fix a few small
semantic problems in the process.

For example:

Indexing a vector by a vector produces a vector nearly all the time.
This CL changes that to all the time, so that programs can have a
consistent expectation about the rank of the result and whether
it can be indexed. For example, suppose we have a vector of values
and want to extract the positive ones, extract those indexes from
a vector v, and then sort v.

	v = v[(x > 0) sel x]
	v = v[up v]

(x > 0) sel x is always a vector. If it has length > 1, then v[(x > 0) sel x]
is always a vector. But if it has length == 1, then before this CL,
v[(x > 0) sel x] was a scalar, not a vector, making the indexing in
v[up v] disallowed.

The shrinking of v in the single-element vector case was important
because it made v[1] be a scalar: the code promoted both v and i
in v[i] to the same overall type, meaning the scalar 1 was promoted
to a vector 1, and then v[1] needed to drop the result back to scalar
to be useful. Taking advantage of the recently added ability to have
different types for the two parts, this CL drops the promotion and
then preserves the form.

There were also a few other lesser surprises arising from promoting
both v and i in v[i] to the same overall type. For example (before this CL):

	1[1]

	1[1 2 3]

Now both of those are diagnosed as erroneous.

These semantic problems are now tested, and the cleanup and changes
did not affect any existing test, suggesting that all the important
behaviors have been preserved.