ivy: make scan linear for associative operators
Scan is in general a quadratic operation, but associative operators
(arguably the most useful ones for a scan) can run in linear time instead
by working left-to-right and reusing results, instead of right-to-left.

For example:

	+\ iota 100000

runs in a small fraction of a second left-to-right
but takes about 10 minutes right-to-left.

Work left-to-right for known associative operators for speed.