ivy: avoid use of semantically-significant trailing whitespace in tests
Many editors today are configured to trim trailing spaces from files,
a kind of 'txtfmt' applied to all files to avoid spurious diffs when editing code.
The ivy test results depend on having tab-indented blank lines,
which are difficult to preserve in those editors (and difficult to see in even more).

This commit changes the test reader to treat blank lines followed by
indented lines as part of output, and blank lines followed by
unindented lines as part of input, removing the need for
tab-indented blank lines.

Some test outputs check for a leading blank line, which cannot be
expressed directly using this rule. For those, any golden output line
consisting of a single "#" is read as a blank line.