ivy: optimize away some Int -> Value allocations
These cases are particularly common when operating on
large boolean matrices. Avoid significant amounts of garbage
allocating zeros and ones.

Commit ded8201e changed zero and one from Value to Int
saying "No reason not to, and it takes a few bytes out of the binary".
The reason not to is to avoid allocations, so add a comment
to that effect.