config: make the formatting of the )cpu special command a little nicer
I found the original a little hard to read, as the spaces were in the wrong places.
Now it looks like this,

4.329ms (3.945ms user, 638µs sys)

which means the new info is parenthesized and feels helpful rather
than confusing.
Plus, particularly for microseconds, it was common to see all zeros
in the decimals, so I dropped them in that case.