ivy: fix bug in expression evaluation order within a vector
When evaluating the elements of a vector, we need to proceed
right to left, as always. This matters when we do things like
	(x+1) (x=2)

The old code did all the assignments first, which was one bug,
and then proceeded left to right, which was another. Delete
the first loop (assigment) and do the other (evaluation)
right to left. Kills two birds with one loop.

Fixes https://github.com/robpike/ivy/issues/95