ivy: statically distinguish local from global variables (#105)
Following Python's lead, a local variable is one that is written
before being read in the program execution order, and
a global variable is one that is read before being written.

The statement 'x;' serves to declare x as a global, assuming
x has not been assigned yet. We could always define an operator too.

Because we have this analysis, we can also count the number
of local variables, assign them indices, and use a slice instead of a map
to hold them, so this change does that as well.

Fixes #82.