ivy: add hyperbolic functions, fix documentation for trig
Add sinh, cosh, tanh as unary functions.

In adding the documentation, I discovered that the way sin etc. are
documented is mishandled, a consequence of APL making them binary
for no good reason, and some code in helpgen.go trying to work
around that.

Instead we just put them all in the unary table, where they belong,
and delete the unnecessary code in helpgen.go. Now

)help atan

works, which it did not before.