value: handle domain better for asin and acos
If the argument is real but out of the range -1 to 1, the function
is still defined. We got this wrong.

Before this change, we just got a confusing error; now we compute
the result from the complex plane:

	asin 2
	1.57079632679j-1.31695789692

Also clean up the constants a bit. More to do there.

Also is it worth making sure we are consistent about which sheet
we return the value from when the function is multivalued?