scan: fix error return in acceptNumber
As a consequence of restructuring when adding complex numbers and tidying
up the scanner, a few places in the scanner for numbers did
	l.errorf(...)
	return false, lexAny
but l.errorf returns the state function itself. As a result, bad numbers such as
	1_0
would not produce an error, and leave the scanner in a funny state.

Fix the obvious way: return the value of the call to l.errorf.