ivy: allow for placeholder indexes, standing for the dimension at that point
If we want to extract a block inside a matrix, say the 2nd column
of a 3x3 matrix, we can write

	x[iota 3; 2]

but one must know the dimension ahead of time. When programming,
one can use rho to pull this out, but it's messy.

This change provides a shorthand: an empty index expression evaluates
to iota "dimension here". Thus the above can be shortened to

	x[;2]

Empty components can be in any location, so

	x[;;]

is the same as x in our example. (So is x[;], but that's a little
more subtle.) And because of how ivy works already

	x[2;]

is the same as

	x[2]

On the drive by, allow an origin to be any non-negative value.