ivy: allow user to specify whether 1/3 is a single token or a division operation
Although ivy was created to do exact rational arithmetic, it has many uses
and sometimes the peculiar handling of / gets in the way. Add a configuration
option and command-line flag to allow one to disable the special handling.
By default,
	1/3
is a single token, rational 1/3. By flipping the flag with the scandivision
special command or the -div command-line flag, it becomes a plain division.
This affects scanning:
	1/2/3
is (1/2)/3 == 1/6 by default, or with the flag flipped, treating / as a division
operator always and parsing right to left, 1/(2/3) == 3/2.

Fixes #120