ivy: make the output of unary rho always be a vector
This was an oversight from the beginning. The unary rho
operator should always return a vector, so for a scalar
the result is an empty vector. Then
	rho rho x
tells the dimensionality of x:
	0: 	scalar
	1:	vector
	2...:	matrix

This is an old bug, and fixing it will make certain programming
problems easier; without it rho of an empty vector gave the
same result as rho of a scalar, which is obviously wrong.

A few tests needed updating.

This may break some scripts, but doesn't affect any permanent
ones of mine.