ivy: better formatting for vectors containing multiline elements
Make the columns and rows line up in a meaningful way when there
is, for example, a 2d matrix in the middle of a vector.  There are
surely many cases this won't handle well, but they should be quite
rare, such as a 5d matrix inside a vector. And for the common cases
that might actually arise, the output is reasonable.

Also add testdata/print.ivy to have a single place to write
tests for this. There may be many as things evolve.

The same treatment for matrices will come next, or soon at least.

Update #117