ivy: improve matrix printing when elements are compound
Building on the work done in the previous commit to make vector
printing handle multiline objects well, provide the same service
for matrices. This step was arguably simpler: all we need to do is
treat the matrix as a vector, print it, and then cut and paste the
pieces together. It works well and looks as well as it is likely
to.

x= 3 4 5 rho iota 100; x[2;2;2]=3 4 rho iota 12; x
            1             2             3             4             5
            6             7             8             9            10
           11            12            13            14            15
           16            17            18            19            20

           21            22            23            24            25
           26 ( 1  2  3  4|            28            29            30
              | 5  6  7  8|
              | 9 10 11 12)
           31            32            33            34            35
           36            37            38            39            40

           41            42            43            44            45
           46            47            48            49            50
           51            52            53            54            55
           56            57            58            59            60

Fixes #117