ivy: add count unary operator
Also known as tally in some APLs, this returns as a scalar the
top-level size of the object: 1 for scalars, length for vectors,
and the first dimension of a matrix.

Also fix a bug in the help code. The docs acquired Go comment
markers at some point and this broke the ability to show some
text using )help.