ivy: add reduce and scan along first axis
One problem is notation, and after years of consideration I came up with
	/%
because the percent sign is otherwise unused (and this doesn't necessarily
steal it from further use), and looks a bit like a flip, as in the fact that
	op/% x
is the same as
	transp op/ transp x

Scan's bookkeeping is so tricky that I implemented with a double
transpose, which is still linear in the size of the matrix so isn't
punitive. I also put in simplified code for the common 2d case.
It works 3 times or more faster for medium sized matrices,
like 1000x1000, less so (but still measurable) for enormous
ones, which are probably limited by memory pressure.