ivy: add fill elements for large takes
This turned out much trickier than I expected. The existing code
was written by rsc and is efficient, but after a few hours of
fiddling I was unable to find a way to make it manage fill items
as well.  So I wrote a simple iterative algorithm that just computes
the geometry for each element, clipping as required.  Good enough
for now: it's still O(n), just with a (much) bigger constant. A
faster method can surely be found, but now we have a good test suite
to verify it if we do.

I tested that my code handles all the old situations as well as the
new ones, but then turned it off unless fills were needed, for
performance.

Update #131