ivy: modulo and encode for non-integers
Previously encode, which calculates modulus, only worked for integers
because only math/big.Ints make it easy. But now we do the full
job, which is a fun exercise.

The big question is what to do with negative values. According to
the APL wiki, there is no consensus about what to do in this case.
So, following precedence set by math/big.Int, and also what seems
right to me, we use the definition set by

	Raymond T. Boute, “The Euclidean definition of the functions
	div and mod”. ACM Transactions on Programming Languages
	and Systems (TOPLAS)

which is what math/big.int.DivMod does. Other than odd cases like
negabinary (which this won't handle), the decision probably doesn't
matter much.

Along the way, we get to implement the "mod" operator for rationals
and floats as well. And with this new structure and understanding,
"encode" becomes a bit cleaner.

Fixes #136