ivy/testdata: add another test for complex floor
The result is surprising but "correct", in agreement with J and
McDonnell's paper but not with Dyalog. Apprently Dyalog uses > in
their equivalent of this line:

    if isTrue("floor", ctx.EvalBinary(x, ">=", y)) {

In my opinion, which no one cares about, the diagonal biase should
be treated like rounding in floating point and go in alternate
directions according to (x^x)^1.