value: add BigInt.Float64 for Go 1.20 and older
In Go 1.21, big.Int got a new method Float64 (proposal golang/go#56984).
In Go 1.21, BigInt has that method because it embeds *big.Int.
In Go 1.20 and older, it's possible to implement it ourselves.

The README and go.mod say that ivy requires Go 1.17 to be built, so
add the Float64 implementation with an appropriate build constraint.