ivy: add mix and split
These operators, inverses (unless fill is involved in mix), turn a
nested matrix into a vector and vice versa:

	mix (1 2) (3 4)
	1 2
	3 4

	split _
	(1 2) (3 4)

In mix, if the elements of the vector are of different shapes, each
is filled out on all axes to the maximum size of any element on
that axis. This is accomplished by using the take operator internally.

 	mix (2 5 rho iota 10) (3 4 rho iota 12)
	 1  2  3  4  5
	 6  7  8  9 10
	 0  0  0  0  0

	 1  2  3  4  0
	 5  6  7  8  0
	 9 10 11 12  0