ivy: fix a bug in operator parsing The definition of a user-defined operator is terminated by a blank line. I forgot about comments. Prior to this fix, the definition
op f arg =
	# return arg
	arg

draws a syntax error because the comment line is lexed into a simple
newline, creating a blank line. The fix is simple: include the
comment leading up to the newline itself in the Newline token.

Note that a line that is only spaces will still count as a blank line,
because it's still blank, and this feels right.