ivy: clean up time zone management
Turn it into location management, that is, track locations rather than
time zones. The code is more robust now, and this fixes the output
for the zero of Unix time in New York. However, there is still some
thinking to do about time zones vs. locations in the UI. More may
come, such as changing the special command to ")location".

Update #158