scan: don't read past newline in peek2
This fixes a buglet in things like help. Before this fix,

)help +

required an extra newline before it would return, as the '+'
led to a call to peek2, which would then read two runes,
the newline and whatever's next. Since peek2 is only called
when building a operator, if the first rune is a newline there
is no point in reading another rune, so don't bother to do so.