ivy: move expr types from parse to value (#186)
It simplifies the function implementation and sets us up better for
vector arguments to allow calling what is currently parse.assignment
in exec.Function.EvalUnary/EvalBinary. But parse imports exec for Function,
so exec cannot import parse for an exported version of assignment.

Instead, move the expression types from parse to value, next to Expr,
along with the assignment code.