ivy: add === and !== operators for non-elementwise comparison
This has been brewing for a while. Now that OrderedCompare is in,
it is easy to provide the approximate equivalent of APL Match and
Not Match, here done with === and !==. These operators look at the
operands as a whole, not elementwise, even when they have very
different types. The functionality was already there for set
operations, this change just makes them available directly to the
user.