ivy: fix )op print when there are vector arguments
The code for printing ops, which is used in save, did not expect
ops defined with vector arguments. Fix that. While we're at it,
use a strings.Builder to make it a little more efficient, although
it will take quite a large op defintion to make the difference
noticeable.

Fixes #196