ivy: add where, unary operator to find indexes of matching elements
This is the APL operator ⍸, a.k.a. where. It returns a vector of
indexes of the locations of non-zero elements in the argument.

Note that the if the argument is a matrix, the result is vector of
vectors, which cannot be used directly as indexes in the current
implementation of indexing. This deserves attention and I'll file
an issue for it.

Fixes #192