ivy: make vector data immutable, copy-on-write (#204)
This "benchmark" does 10,000 function calls passing
an ignored 1000x1000 matrix as an argument.
The old implementation made a copy every time a value
was assigned to a variable, making the parameter passing
very expensive. The implementation in this CL makes a
copy each time modified data is copied out of a variable,
eliminating all the copies when the data has not changed
(and eliminating the Value.Copy method).

% time ivy0.3.11 testdata/copy.ivy
1000 1000
1000 1000
       58.22 real        93.94 user        42.62 sys
% time ivy testdata/copy.ivy
1000 1000
1000 1000
        0.27 real         0.04 user         0.01 sys
%

The main change here is that Vector can no longer be used
for any mutable operations. Everything must use vectorEditor instead.
To enforce this, Vector's All and Slice methods now return iterators
over the data, making it impossible to mutate the data,
and the Writable method is gone.
However, Vector now has an edit method that returns a vectorEditor,
which edits a copy of the Vector. Changes can be made using the
vectorEditor, which has the usual Vector methods as well as
Set(index, value) and Append(values...), and then converted back
to a Vector by calling the Publish method.

A follow-up change may replace the underlying storage
with a more fine-grained copy-on-write tree data structure,
but this is already a huge improvement, and it sets up for
alternate representations in the Vector and vectorEditor types.
Changing the underlying data structure will now only require
changing the Vector and vectorEditor implementations.