ivy: implement index by a vector (or matrix) of vectors
The new 'where' operator returns a vector of vectors if the argument is a matrix:

	where 2 != 2 2 rho iota 4
		(1 1) (2 1) (2 2)

The intent is that those tuples can be used to index the matrix. Make it so.

	m[where (m=5 5 rho iota 25) in 2 3 5 7 11]
		2 3 5 7 11

	4 5 rho 'abcdefghijklmnopqrstuvwxyz')[iota 2 3]
		abc
		fgh

Indexing a matrix by a vector or matrix of vectors
is the same as indexing a vector by a vector or matrix of scalars,
except for the actual indexes being vectors instead of scalars.
The vector lengths must be uniform.
If the vector length is smaller than the rank of the matrix,
whole rows or submatrices are selected.

	(5 5 5 rho iota 125)[(1 2) (3 4)]
		 6  7  8  9 10
		66 67 68 69 70

Assignment works the same way:

	m=5 5 rho iota 25
	m[where m in 2 3 5 7 11] = - 2 3 5 7 11
	m
		  1  -2  -3   4  -5
		  6  -7   8   9  10
		-11  12  13  14  15
		 16  17  18  19  20
		 21  22  23  24  25

This change also fixes an order of evaluation bug where
in m[x;y;z], a variable m was being evaluated before z, y, and x.
If z, y, or x defined m (as in the second example above),
ivy incorrectly reported an undefined global.

Fixes #207.